[dotnet-linker] Move methods from ManagedRegistrarLookupTablesStep to AppBundleRewriter.
Move FindNSObjectConstructor, FindINativeObjectConstructor,
ImplementConstructNSObjectFactoryMethod, ImplementConstructINativeObjectFactoryMethod,
and AddTypeInterfaceImplementation from ManagedRegistrarLookupTablesStep to
AppBundleRewriter, and update all call sites accordingly.
