+++ tools/dotnet-linker/AppBundleRewriter.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Text;
@@ -1725,6 +1726,176 @@ namespace Xamarin.Linker {
 			created = true;
 
 			return rv;
+		}
+
+		internal static MethodDefinition? FindNSObjectConstructor (TypeDefinition type)
+		{
+			return FindConstructorWithOneParameter ("ObjCRuntime", "NativeHandle")
+				?? FindConstructorWithOneParameter ("System", "IntPtr");
+
+			MethodDefinition? FindConstructorWithOneParameter (string ns, string cls)
+				=> type.Methods.SingleOrDefault (method =>
+					method.IsConstructor
+						&& !method.IsStatic
+						&& method.HasParameters
+						&& method.Parameters.Count == 1
+						&& method.Parameters [0].ParameterType.Is (ns, cls));
+		}
+
+		internal static MethodDefinition? FindINativeObjectConstructor (TypeDefinition type)
+		{
+			return FindConstructorWithTwoParameters ("ObjCRuntime", "NativeHandle", "System", "Boolean")
+				?? FindConstructorWithTwoParameters ("System", "IntPtr", "System", "Boolean");
+
+			MethodDefinition? FindConstructorWithTwoParameters (string ns1, string cls1, string ns2, string cls2)
+				=> type.Methods.SingleOrDefault (method =>
+					method.IsConstructor
+						&& !method.IsStatic
+						&& method.HasParameters
+						&& method.Parameters.Count == 2
+						&& method.Parameters [0].ParameterType.Is (ns1, cls1)
+						&& method.Parameters [1].ParameterType.Is (ns2, cls2));
+		}
+
+		internal void ImplementConstructNSObjectFactoryMethod (Tuner.DerivedLinkContext context, TypeDefinition type, MethodReference ctor)
+		{
+			var abr = this;
+
+			// skip creating the factory for NSObject itself
+			if (type.Is ("Foundation", "NSObject"))
+				return;
+
+			// Make sure the type implements INSObjectFactory, otherwise we can't override the _Xamarin_ConstructNSObject method from it.
+			AddTypeInterfaceImplementation (abr, context, type, abr.Foundation_INSObjectFactory);
+
+			var createInstanceMethod = type.AddMethod ("_Xamarin_ConstructNSObject", MethodAttributes.Public | MethodAttributes.Static | MethodAttributes.NewSlot | MethodAttributes.HideBySig, abr.Foundation_NSObject);
+			var nativeHandleParameter = createInstanceMethod.AddParameter ("nativeHandle", abr.ObjCRuntime_NativeHandle);
+			abr.Foundation_INSObjectFactory.Resolve ().IsPublic = true;
+			createInstanceMethod.Overrides.Add (abr.INSObjectFactory__Xamarin_ConstructNSObject);
+			var body = createInstanceMethod.CreateBody (out var il);
+
+			if (type.HasGenericParameters) {
+				ctor = type.CreateMethodReferenceOnGenericType (ctor, type.GenericParameters.ToArray ());
+			}
+
+			// return new TypeA (nativeHandle); // for NativeHandle ctor
+			// return new TypeA ((IntPtr) nativeHandle); // for IntPtr ctor
+			il.Emit (OpCodes.Ldarg, nativeHandleParameter);
+			if (ctor.Parameters [0].ParameterType.Is ("System", "IntPtr"))
+				il.Emit (OpCodes.Call, abr.NativeObject_op_Implicit_IntPtr);
+			il.Emit (OpCodes.Newobj, ctor);
+			il.Emit (OpCodes.Ret);
+
+			body.GenerateILOffsets ();
+
+			// make sure the trimmer doesn't trim it away if the type is kept
+			if (context.App.Registrar == RegistrarMode.TrimmableStatic) {
+				// TODO: need to investigate why this is needed (https://github.com/dotnet/macios/issues/25232)
+				abr.AddDynamicDependencyAttributeToStaticConstructor (type, createInstanceMethod);
+			} else {
+				context.Annotations.Mark (createInstanceMethod);
+			}
+		}
+
+		internal void ImplementConstructINativeObjectFactoryMethod (Tuner.DerivedLinkContext context, TypeDefinition type, MethodReference? ctor)
+		{
+			var abr = this;
+
+			// skip creating the factory for NSObject itself
+			if (type.Is ("Foundation", "NSObject"))
+				return;
+
+			// If the type is a subclass of NSObject, we prefer the NSObject "IntPtr" constructor
+			MethodReference? nsobjectConstructor = type.IsNSObject (context) ? AppBundleRewriter.FindNSObjectConstructor (type) : null;
+			if (nsobjectConstructor is null && ctor is null)
+				return;
+
+			// Make sure the type implements INativeObject, otherwise we can't override the _Xamarin_ConstructINativeObject method from it.
+			AddTypeInterfaceImplementation (abr, context, type, abr.ObjCRuntime_INativeObject);
+
+			var createInstanceMethod = type.AddMethod ("_Xamarin_ConstructINativeObject", MethodAttributes.Public | MethodAttributes.Static | MethodAttributes.NewSlot | MethodAttributes.HideBySig, abr.ObjCRuntime_INativeObject);
+			var nativeHandleParameter = createInstanceMethod.AddParameter ("nativeHandle", abr.ObjCRuntime_NativeHandle);
+			var ownsParameter = createInstanceMethod.AddParameter ("owns", abr.System_Boolean);
+			abr.INativeObject__Xamarin_ConstructINativeObject.Resolve ().IsPublic = true;
+			createInstanceMethod.Overrides.Add (abr.INativeObject__Xamarin_ConstructINativeObject);
+			var body = createInstanceMethod.CreateBody (out var il);
+
+			if (nsobjectConstructor is not null) {
+				// var instance = new TypeA (nativeHandle);
+				// // alternatively with a cast: new TypeA ((IntPtr) nativeHandle);
+				// if (instance is not null && owns)
+				//     Runtime.TryReleaseINativeObject (instance);
+				// return instance;
+
+				if (type.HasGenericParameters) {
+					nsobjectConstructor = type.CreateMethodReferenceOnGenericType (nsobjectConstructor, type.GenericParameters.ToArray ());
+				}
+
+				il.Emit (OpCodes.Ldarg, nativeHandleParameter);
+				if (nsobjectConstructor.Parameters [0].ParameterType.Is ("System", "IntPtr"))
+					il.Emit (OpCodes.Call, abr.NativeObject_op_Implicit_IntPtr);
+				il.Emit (OpCodes.Newobj, nsobjectConstructor);
+
+				var falseTarget = il.Create (OpCodes.Nop);
+				il.Emit (OpCodes.Dup);
+				il.Emit (OpCodes.Ldnull);
+				il.Emit (OpCodes.Cgt_Un);
+				il.Emit (OpCodes.Ldarg, ownsParameter);
+				il.Emit (OpCodes.And);
+				il.Emit (OpCodes.Brfalse_S, falseTarget);
+
+				il.Emit (OpCodes.Dup);
+				il.Emit (OpCodes.Call, abr.Runtime_TryReleaseINativeObject);
+
+				il.Append (falseTarget);
+
+				il.Emit (OpCodes.Ret);
+			} else if (ctor is not null) {
+				// return new TypeA (nativeHandle, owns); // for NativeHandle ctor
+				// return new TypeA ((IntPtr) nativeHandle, owns); // IntPtr ctor
+
+				if (type.HasGenericParameters) {
+					ctor = type.CreateMethodReferenceOnGenericType (ctor, type.GenericParameters.ToArray ());
+				}
+
+				il.Emit (OpCodes.Ldarg, nativeHandleParameter);
+				if (ctor.Parameters [0].ParameterType.Is ("System", "IntPtr"))
+					il.Emit (OpCodes.Call, abr.NativeObject_op_Implicit_IntPtr);
+				il.Emit (OpCodes.Ldarg, ownsParameter);
+				il.Emit (OpCodes.Newobj, ctor);
+				il.Emit (OpCodes.Ret);
+			} else {
+				throw new UnreachableException ();
+			}
+
+			body.GenerateILOffsets ();
+
+			// make sure the trimmer doesn't trim it away if the type is kept
+			if (context.App.Registrar == RegistrarMode.TrimmableStatic) {
+				// TODO: need to investigate why this is needed (https://github.com/dotnet/macios/issues/25232)
+				abr.AddDynamicDependencyAttributeToStaticConstructor (type, createInstanceMethod);
+			} else {
+				context.Annotations.Mark (createInstanceMethod);
+			}
+		}
+
+		static void AddTypeInterfaceImplementation (AppBundleRewriter abr, Tuner.DerivedLinkContext context, TypeDefinition type, TypeReference iface)
+		{
+			if (type.HasInterfaces && type.Interfaces.Any (v => v.InterfaceType == iface))
+				return;
+
+			var ifaceImplementation = new InterfaceImplementation (iface);
+			type.Interfaces.Add (ifaceImplementation);
+
+			// make sure the trimmer doesn't trim it away if the type is kept
+			if (context.App.Registrar == RegistrarMode.TrimmableStatic) {
+				// TODO: need to investigate why this is needed (https://github.com/dotnet/macios/issues/25232)
+				abr.AddAttributeToStaticConstructor (type, abr.CreateDynamicDependencyAttribute (DynamicallyAccessedMemberTypes.Interfaces, type));
+			} else {
+				context.Annotations.Mark (ifaceImplementation);
+				context.Annotations.Mark (ifaceImplementation.InterfaceType);
+				context.Annotations.Mark (ifaceImplementation.InterfaceType.Resolve ());
+			}
 		}
 	}
 }
